+++ packages/react/src/primitives/hooks.ts
@@ -1,0 +1,1 @@
+export { usePagination } from './Pagination';
