+++ packages/react/src/primitives/index.tsx
@@ -1,43 +1,5 @@
-export * from './Alert';
-export * from './Badge';
-export * from './Button';
-export * from './ButtonGroup';
-export * from './Card';
-export * from './CheckboxField';
-export * from './Collection';
-export * from './Divider';
-export * from './Expander';
-export * from './FieldGroup';
-export * from './FieldGroupIcon';
-export * from './Flex';
-export * from './Grid';
-export * from './Heading';
-export * from './Icon';
-export * from './Image';
-export * from './Link';
-export * from './Loader';
-export * from './Menu';
-export * from './Pagination';
-export * from './Placeholder';
-export * from './PasswordField';
-export * from './PhoneNumberField';
-export * from './Placeholder';
-export * from './Radio';
-export * from './RadioGroupField';
-export * from './Rating';
-export * from './ScrollView';
-export * from './SearchField';
-export * from './SelectField';
-export * from './SliderField';
-export * from './StepperField';
-export * from './SwitchField';
-export * from './Tabs';
-export * from './Text';
-export * from './TextField';
-export * from './ToggleButton';
-export * from './ToggleButtonGroup';
-export * from './View';
-export * from './VisuallyHidden';
+export * from './hooks';
+export * from './primitives';
 
 export * from './shared';
 export * from './types';
